requirements.txt
streamlit
python-dotenv
peewee
psycopg2-binary